+++ cognite_toolkit/_cdf_tk/rules/__init__.py
@@ -1,7 +1,6 @@
 from ._agents import AgentRuleSet
 from ._auth import CheckDataSetMissing
 from ._base import ToolkitGlobalRuleSet, ToolkitLocalRule
-from ._data_modeling import DataModelingChangeRuleSet
 from ._dependencies import DependencyRuleSet
 from ._functions import FunctionRuleSet
 from ._infield import InFieldCDMRuleSet
@@ -11,7 +10,6 @@ from ._orchestrator import LocalRulesOrchestrator, get_global_rules_registry
 __all__ = [
     "AgentRuleSet",
     "CheckDataSetMissing",
-    "DataModelingChangeRuleSet",
     "DependencyRuleSet",
     "FunctionRuleSet",
     "InFieldCDMRuleSet",
